--- a/Cargo.toml
+++ b/Cargo.toml
@@ -17,5 +17,10 @@ std = []
 serde = { version = "1.0", optional = true }
 
 [dev-dependencies]
+criterion = { version = "0.5" }
 test-case = { version = "3" }
 serde_json = { version = "1" }
+
+[[bench]]
+name = "bench"
+harness = false