Instead of cloning the iterator at each iteration, `N` iterators are initialized when constructed and are iterated over.

```
iterate/ab              time:   [964.52 ps 972.04 ps 980.79 ps]
                        change: [-81.682% -81.501% -81.302%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) low mild
  1 (1.00%) high mild
  4 (4.00%) high severe

iterate/whatever        time:   [9.2355 ns 9.2491 ns 9.2626 ns]
                        change: [-78.668% -78.556% -78.435%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) low mild
  1 (1.00%) high severe

iterate/hello world     time:   [15.009 ns 15.031 ns 15.053 ns]
                        change: [-74.515% -74.348% -74.163%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  4 (4.00%) high severe

iterate/this is a long sentence that will be iterated over
                        time:   [101.77 ns 101.83 ns 101.88 ns]
                        change: [-62.285% -62.036% -61.774%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe
```